chart in separate component
